finalize employee login
